requirements.txt
astroid==2.5.6
autopep8==1.5.7
backcall==0.2.0
click==8.0.1
dataclasses==0.8
decorator==5.0.9
environs==9.3.2
Flask==2.0.1
importlib-metadata==4.5.0
ipdb==0.13.9
ipython==7.16.1
ipython-genutils==0.2.0
isort==5.8.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
marshmallow==3.12.1
mccabe==0.6.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycodestyle==2.7.0
Pygments==2.9.0
pylint==2.8.3
python-dotenv==0.17.1
six==1.16.0
toml==0.10.2
traitlets==4.3.3
typed-ast==1.4.3
typing-extensions==3.10.0.0
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.4.1
